refactor(s2n-quic-core): move MTU controller to core
