+++ quic/s2n-quic-core/src/path/mtu/tests.rs
@@ -1,0 +1,635 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+use super::{testing::*, *};
+use crate::{
+    endpoint,
+    event::testing::Publisher,
+    frame::Frame,
+    packet::number::PacketNumberSpace,
+    recovery::congestion_controller::testing::mock::CongestionController,
+    time::{clock::testing::now, timer::Provider as _},
+    transmission::{
+        writer::{
+            testing::{OutgoingFrameBuffer, Writer as MockWriteContext},
+            Writer as _,
+        },
+        Provider as _,
+    },
+    varint::VarInt,
+};
+use std::{convert::TryInto, net::SocketAddr};
+
+/// Creates an application space packet number with the given value
+fn pn(nr: usize) -> PacketNumber {
+    PacketNumberSpace::ApplicationData.new_packet_number(VarInt::new(nr as u64).unwrap())
+}
+
+#[test]
+fn base_plpmtu_is_1200() {
+    //= https://www.rfc-editor.org/rfc/rfc8899#section-5.1.2
+    //= type=test
+    //# When using
+    //# IPv4, there is no currently equivalent size specified, and a
+    //# default BASE_PLPMTU of 1200 bytes is RECOMMENDED.
+    assert_eq!(BASE_PLPMTU, 1200);
+}
+
+#[test]
+fn min_max_mtu() {
+    // Use an IPv6 address to force a smaller `max_udp_payload`
+    let addr: SocketAddr = "[::1]:123".parse().unwrap();
+    let controller = Controller::new(MaxMtu::MIN, &addr.into());
+    assert_eq!(BASE_PLPMTU, controller.plpmtu);
+}
+
+#[test]
+fn new_max_mtu_smaller_than_common_mtu() {
+    let mut controller = new_controller(BASE_PLPMTU + UDP_HEADER_LEN + IPV4_MIN_HEADER_LEN + 1);
+    assert_eq!(BASE_PLPMTU + 1, controller.probed_size);
+
+    controller.enable();
+    assert_eq!(State::SearchComplete, controller.state);
+}
+
+#[test]
+fn new_ipv4() {
+    let addr: SocketAddr = "127.0.0.1:443".parse().unwrap();
+    let controller = Controller::new(1600.try_into().unwrap(), &addr.into());
+    assert_eq!(
+        1600 - UDP_HEADER_LEN - IPV4_MIN_HEADER_LEN,
+        controller.max_udp_payload
+    );
+    assert_eq!(
+        1600 - UDP_HEADER_LEN - IPV4_MIN_HEADER_LEN,
+        controller.max_probe_size
+    );
+    assert_eq!(BASE_PLPMTU as usize, controller.mtu());
+    assert_eq!(0, controller.probe_count);
+    assert_eq!(State::Disabled, controller.state);
+    assert!(!controller.pmtu_raise_timer.is_armed());
+    assert_eq!(
+        ETHERNET_MTU - UDP_HEADER_LEN - IPV4_MIN_HEADER_LEN,
+        controller.probed_size
+    );
+}
+
+#[test]
+fn new_ipv6() {
+    let addr: SocketAddr = "[2001:0db8:85a3:0001:0002:8a2e:0370:7334]:9000"
+        .parse()
+        .unwrap();
+    let controller = Controller::new(2000.try_into().unwrap(), &addr.into());
+    assert_eq!(
+        2000 - UDP_HEADER_LEN - IPV6_MIN_HEADER_LEN,
+        controller.max_udp_payload
+    );
+    assert_eq!(
+        2000 - UDP_HEADER_LEN - IPV6_MIN_HEADER_LEN,
+        controller.max_probe_size
+    );
+    assert_eq!(BASE_PLPMTU as usize, controller.mtu());
+    assert_eq!(0, controller.probe_count);
+    assert_eq!(State::Disabled, controller.state);
+    assert!(!controller.pmtu_raise_timer.is_armed());
+    assert_eq!(
+        ETHERNET_MTU - UDP_HEADER_LEN - IPV6_MIN_HEADER_LEN,
+        controller.probed_size
+    );
+}
+
+#[test]
+fn enable_already_enabled() {
+    let mut controller = new_controller(1500);
+    assert_eq!(State::Disabled, controller.state);
+    controller.enable();
+    assert_eq!(State::SearchRequested, controller.state);
+    controller.state = State::SearchComplete;
+    controller.enable();
+    assert_eq!(State::SearchComplete, controller.state);
+}
+
+#[test]
+fn enable() {
+    let mut controller = new_controller(1500);
+    assert_eq!(State::Disabled, controller.state);
+    controller.enable();
+    assert_eq!(State::SearchRequested, controller.state);
+}
+
+//= https://www.rfc-editor.org/rfc/rfc8899#section-4.2
+//= type=test
+//# When
+//# supported, this mechanism MAY also be used by DPLPMTUD to acknowledge
+//# reception of a probe packet.
+#[test]
+fn on_packet_ack_within_threshold() {
+    let mut controller = new_controller(1472 + PROBE_THRESHOLD * 2);
+    let max_udp_payload = controller.max_udp_payload;
+    let pn = pn(1);
+    let mut cc = CongestionController::default();
+    let now = now();
+    let mut publisher = Publisher::snapshot();
+    controller.state = State::Searching(pn, now);
+    controller.probed_size = BASE_PLPMTU;
+    controller.max_probe_size = BASE_PLPMTU + PROBE_THRESHOLD * 2 - 1;
+
+    controller.on_packet_ack(
+        pn,
+        BASE_PLPMTU,
+        &mut cc,
+        path::Id::test_id(),
+        &mut publisher,
+    );
+
+    assert_eq!(
+        BASE_PLPMTU + (max_udp_payload - BASE_PLPMTU) / 2,
+        controller.probed_size
+    );
+    assert_eq!(1, cc.on_mtu_update);
+    assert_eq!(State::SearchComplete, controller.state);
+    assert!(controller.pmtu_raise_timer.is_armed());
+    assert_eq!(
+        Some(now + PMTU_RAISE_TIMER_DURATION),
+        controller.next_expiration()
+    );
+
+    // Enough time passes that its time to try raising the PMTU again
+    let now = now + PMTU_RAISE_TIMER_DURATION;
+    controller.on_timeout(now);
+
+    assert_eq!(State::SearchRequested, controller.state);
+    assert_eq!(
+        BASE_PLPMTU + (max_udp_payload - BASE_PLPMTU) / 2,
+        controller.probed_size
+    );
+}
+
+#[test]
+fn on_packet_ack_within_threshold_of_max_plpmtu() {
+    let mut controller = new_controller(1472 + (PROBE_THRESHOLD * 2 - 1));
+    let max_udp_payload = controller.max_udp_payload;
+    let pn = pn(1);
+    let mut cc = CongestionController::default();
+    let now = now();
+    let mut publisher = Publisher::snapshot();
+    controller.state = State::Searching(pn, now);
+
+    controller.on_packet_ack(
+        pn,
+        controller.probed_size,
+        &mut cc,
+        path::Id::test_id(),
+        &mut publisher,
+    );
+
+    assert_eq!(1472 + (max_udp_payload - 1472) / 2, controller.probed_size);
+    assert_eq!(1, cc.on_mtu_update);
+    assert_eq!(State::SearchComplete, controller.state);
+    assert!(!controller.pmtu_raise_timer.is_armed());
+}
+
+//= https://www.rfc-editor.org/rfc/rfc8899#section-5.3.2
+//= type=test
+//# Implementations SHOULD select the set of probe packet sizes to
+//# maximize the gain in PLPMTU from each search step.
+
+//= https://www.rfc-editor.org/rfc/rfc8899#section-8
+//= type=test
+//# To avoid excessive load, the interval between individual probe
+//# packets MUST be at least one RTT, and the interval between rounds of
+//# probing is determined by the PMTU_RAISE_TIMER.
+#[test]
+fn on_packet_ack_search_requested() {
+    let mut controller = new_controller(1500 + (PROBE_THRESHOLD * 2));
+    let max_udp_payload = controller.max_udp_payload;
+    let pn = pn(1);
+    let mut cc = CongestionController::default();
+    let now = now();
+    let mut publisher = Publisher::snapshot();
+    controller.state = State::Searching(pn, now);
+
+    controller.on_packet_ack(
+        pn,
+        controller.probed_size,
+        &mut cc,
+        path::Id::test_id(),
+        &mut publisher,
+    );
+
+    assert_eq!(1472 + (max_udp_payload - 1472) / 2, controller.probed_size);
+    assert_eq!(1, cc.on_mtu_update);
+    assert_eq!(State::SearchRequested, controller.state);
+    assert!(!controller.pmtu_raise_timer.is_armed());
+}
+
+#[test]
+fn on_packet_ack_resets_black_hole_counter() {
+    let mut controller = new_controller(1500 + (PROBE_THRESHOLD * 2));
+    let pnum = pn(3);
+    let mut cc = CongestionController::default();
+    let mut publisher = Publisher::snapshot();
+    controller.enable();
+
+    controller.black_hole_counter += 1;
+    // ack a packet smaller than the plpmtu
+    controller.on_packet_ack(
+        pnum,
+        controller.plpmtu - 1,
+        &mut cc,
+        path::Id::test_id(),
+        &mut publisher,
+    );
+    assert_eq!(controller.black_hole_counter, 1);
+    assert_eq!(None, controller.largest_acked_mtu_sized_packet);
+
+    // ack a packet the size of the plpmtu
+    controller.on_packet_ack(
+        pnum,
+        controller.plpmtu,
+        &mut cc,
+        path::Id::test_id(),
+        &mut publisher,
+    );
+    assert_eq!(controller.black_hole_counter, 0);
+    assert_eq!(Some(pnum), controller.largest_acked_mtu_sized_packet);
+
+    controller.black_hole_counter += 1;
+
+    // ack an older packet
+    let pnum_2 = pn(2);
+    controller.on_packet_ack(
+        pnum_2,
+        controller.plpmtu,
+        &mut cc,
+        path::Id::test_id(),
+        &mut publisher,
+    );
+    assert_eq!(controller.black_hole_counter, 1);
+    assert_eq!(Some(pnum), controller.largest_acked_mtu_sized_packet);
+}
+
+#[test]
+fn on_packet_ack_disabled_controller() {
+    let mut controller = new_controller(1500 + (PROBE_THRESHOLD * 2));
+    let pnum = pn(3);
+    let mut cc = CongestionController::default();
+    let mut publisher = Publisher::snapshot();
+
+    controller.black_hole_counter += 1;
+    controller.largest_acked_mtu_sized_packet = Some(pnum);
+
+    let pn = pn(10);
+    controller.on_packet_ack(
+        pn,
+        controller.plpmtu,
+        &mut cc,
+        path::Id::test_id(),
+        &mut publisher,
+    );
+
+    assert_eq!(State::Disabled, controller.state);
+    assert_eq!(controller.black_hole_counter, 1);
+    assert_eq!(Some(pnum), controller.largest_acked_mtu_sized_packet);
+}
+
+#[test]
+fn on_packet_ack_not_application_space() {
+    let mut controller = new_controller(1500 + (PROBE_THRESHOLD * 2));
+    let pnum = pn(3);
+    let mut cc = CongestionController::default();
+    let mut publisher = Publisher::snapshot();
+    controller.enable();
+
+    controller.black_hole_counter += 1;
+    controller.largest_acked_mtu_sized_packet = Some(pnum);
+
+    // on_packet_ack will be called with packet numbers from Initial and Handshake space,
+    // so it should not fail in this scenario.
+    let pn = PacketNumberSpace::Handshake.new_packet_number(VarInt::from_u8(10));
+    controller.on_packet_ack(
+        pn,
+        controller.plpmtu,
+        &mut cc,
+        path::Id::test_id(),
+        &mut publisher,
+    );
+    assert_eq!(controller.black_hole_counter, 1);
+    assert_eq!(Some(pnum), controller.largest_acked_mtu_sized_packet);
+}
+
+//= https://www.rfc-editor.org/rfc/rfc8899#section-3
+//= type=test
+//# The PL is REQUIRED to be
+//# robust in the case where probe packets are lost due to other
+//# reasons (including link transmission error, congestion).
+#[test]
+fn on_packet_loss() {
+    let mut controller = new_controller(1500);
+    let max_udp_payload = controller.max_udp_payload;
+    let pn = pn(1);
+    let mut cc = CongestionController::default();
+    let now = now();
+    let mut publisher = Publisher::snapshot();
+    controller.state = State::Searching(pn, now);
+    let probed_size = controller.probed_size;
+
+    controller.on_packet_loss(
+        pn,
+        controller.probed_size,
+        false,
+        now,
+        &mut cc,
+        path::Id::test_id(),
+        &mut publisher,
+    );
+
+    assert_eq!(0, cc.on_mtu_update);
+    assert_eq!(max_udp_payload, controller.max_probe_size);
+    assert_eq!(probed_size, controller.probed_size);
+    assert_eq!(State::SearchRequested, controller.state);
+}
+
+#[test]
+fn on_packet_loss_max_probes() {
+    let mut controller = new_controller(1500);
+    let max_udp_payload = controller.max_udp_payload;
+    let pn = pn(1);
+    let mut cc = CongestionController::default();
+    let now = now();
+    let mut publisher = Publisher::snapshot();
+    controller.state = State::Searching(pn, now);
+    controller.probe_count = MAX_PROBES;
+    assert_eq!(max_udp_payload, controller.max_probe_size);
+
+    controller.on_packet_loss(
+        pn,
+        controller.probed_size,
+        false,
+        now,
+        &mut cc,
+        path::Id::test_id(),
+        &mut publisher,
+    );
+
+    assert_eq!(0, cc.on_mtu_update);
+    assert_eq!(1472, controller.max_probe_size);
+    assert_eq!(
+        BASE_PLPMTU + (1472 - BASE_PLPMTU) / 2,
+        controller.probed_size
+    );
+    assert_eq!(State::SearchRequested, controller.state);
+}
+
+#[test]
+fn on_packet_loss_black_hole() {
+    let mut controller = new_controller(1500);
+    let mut cc = CongestionController::default();
+    let now = now();
+    let mut publisher = Publisher::snapshot();
+    controller.plpmtu = 1472;
+    controller.enable();
+
+    for i in 0..BLACK_HOLE_THRESHOLD + 1 {
+        let pn = pn(i as usize);
+
+        // Losing a packet the size of the BASE_PLPMTU should not increase the black_hole_counter
+        controller.on_packet_loss(
+            pn,
+            BASE_PLPMTU,
+            true,
+            now,
+            &mut cc,
+            path::Id::test_id(),
+            &mut publisher,
+        );
+        assert_eq!(controller.black_hole_counter, i);
+
+        // Losing a packet larger than the PLPMTU should not increase the black_hole_counter
+        controller.on_packet_loss(
+            pn,
+            controller.plpmtu + 1,
+            true,
+            now,
+            &mut cc,
+            path::Id::test_id(),
+            &mut publisher,
+        );
+        assert_eq!(controller.black_hole_counter, i);
+
+        // Losing a packet that does not start a new loss burst should not increase the black_hole_counter
+        controller.on_packet_loss(
+            pn,
+            BASE_PLPMTU + 1,
+            false,
+            now,
+            &mut cc,
+            path::Id::test_id(),
+            &mut publisher,
+        );
+        assert_eq!(controller.black_hole_counter, i);
+
+        controller.on_packet_loss(
+            pn,
+            BASE_PLPMTU + 1,
+            true,
+            now,
+            &mut cc,
+            path::Id::test_id(),
+            &mut publisher,
+        );
+        if i < BLACK_HOLE_THRESHOLD {
+            assert_eq!(controller.black_hole_counter, i + 1);
+        }
+    }
+
+    assert_eq!(controller.black_hole_counter, 0);
+    assert_eq!(None, controller.largest_acked_mtu_sized_packet);
+    assert_eq!(1, cc.on_mtu_update);
+    assert_eq!(BASE_PLPMTU, controller.plpmtu);
+    assert_eq!(State::SearchComplete, controller.state);
+    assert_eq!(
+        Some(now + BLACK_HOLE_COOL_OFF_DURATION),
+        controller.pmtu_raise_timer.next_expiration()
+    );
+}
+
+#[test]
+fn on_packet_loss_disabled_controller() {
+    let mut controller = new_controller(1500);
+    let mut cc = CongestionController::default();
+    let now = now();
+    let mut publisher = Publisher::snapshot();
+
+    for i in 0..BLACK_HOLE_THRESHOLD + 1 {
+        let pn = pn(i as usize);
+        assert_eq!(controller.black_hole_counter, 0);
+        controller.on_packet_loss(
+            pn,
+            BASE_PLPMTU + 1,
+            false,
+            now,
+            &mut cc,
+            path::Id::test_id(),
+            &mut publisher,
+        );
+    }
+
+    assert_eq!(State::Disabled, controller.state);
+    assert_eq!(controller.black_hole_counter, 0);
+    assert_eq!(0, cc.on_mtu_update);
+}
+
+#[test]
+fn on_packet_loss_not_application_space() {
+    let mut controller = new_controller(1500);
+    let mut cc = CongestionController::default();
+    let mut publisher = Publisher::snapshot();
+
+    // test the loss in each state
+    for state in [
+        State::Disabled,
+        State::SearchRequested,
+        State::Searching(pn(1), now()),
+        State::SearchComplete,
+    ] {
+        controller.state = state;
+        for i in 0..BLACK_HOLE_THRESHOLD + 1 {
+            // on_packet_loss may be called with packet numbers from Initial and Handshake space
+            // so it should not fail in this scenario.
+            let pn = PacketNumberSpace::Initial.new_packet_number(VarInt::from_u8(i));
+            controller.on_packet_loss(
+                pn,
+                BASE_PLPMTU + 1,
+                false,
+                now(),
+                &mut cc,
+                path::Id::test_id(),
+                &mut publisher,
+            );
+            assert_eq!(controller.black_hole_counter, 0);
+            assert_eq!(0, cc.on_mtu_update);
+        }
+    }
+}
+
+//= https://www.rfc-editor.org/rfc/rfc8899#section-5.2
+//= type=test
+//# When used with an
+//# acknowledged PL (e.g., SCTP), DPLPMTUD SHOULD NOT continue to
+//# generate PLPMTU probes in this state.
+#[test]
+fn on_transmit_search_not_requested() {
+    let mut controller = new_controller(1500);
+    controller.state = State::SearchComplete;
+    let mut frame_buffer = OutgoingFrameBuffer::new();
+    let mut write_context = MockWriteContext::new(
+        now(),
+        &mut frame_buffer,
+        transmission::Constraint::None,
+        transmission::Mode::MtuProbing,
+        endpoint::Type::Server,
+    );
+
+    controller.on_transmit(&mut write_context);
+    assert!(frame_buffer.is_empty());
+    assert_eq!(State::SearchComplete, controller.state);
+}
+
+#[test]
+fn on_transmit_not_mtu_probing() {
+    let mut controller = new_controller(1500);
+    controller.state = State::SearchRequested;
+    let mut frame_buffer = OutgoingFrameBuffer::new();
+    let mut write_context = MockWriteContext::new(
+        now(),
+        &mut frame_buffer,
+        transmission::Constraint::None,
+        transmission::Mode::Normal,
+        endpoint::Type::Server,
+    );
+
+    controller.on_transmit(&mut write_context);
+    assert!(frame_buffer.is_empty());
+    assert_eq!(State::SearchRequested, controller.state);
+
+    controller.state = State::SearchComplete;
+    let mut frame_buffer = OutgoingFrameBuffer::new();
+    let mut write_context = MockWriteContext::new(
+        now(),
+        &mut frame_buffer,
+        transmission::Constraint::None,
+        transmission::Mode::Normal,
+        endpoint::Type::Server,
+    );
+
+    controller.on_transmit(&mut write_context);
+    assert!(frame_buffer.is_empty());
+    assert_eq!(State::SearchComplete, controller.state);
+}
+
+#[test]
+fn on_transmit_no_capacity() {
+    let mut controller = new_controller(1500);
+    controller.state = State::SearchRequested;
+    let mut frame_buffer = OutgoingFrameBuffer::new();
+    frame_buffer.set_max_packet_size(Some(controller.probed_size as usize - 1));
+    let mut write_context = MockWriteContext::new(
+        now(),
+        &mut frame_buffer,
+        transmission::Constraint::None,
+        transmission::Mode::MtuProbing,
+        endpoint::Type::Server,
+    );
+
+    controller.on_transmit(&mut write_context);
+    assert!(frame_buffer.is_empty());
+    assert_eq!(State::SearchComplete, controller.state);
+}
+
+//= https://www.rfc-editor.org/rfc/rfc9000#section-14.4
+//= type=test
+//# Endpoints could limit the content of PMTU probes to PING and PADDING
+//# frames, since packets that are larger than the current maximum
+//# datagram size are more likely to be dropped by the network.
+
+//= https://www.rfc-editor.org/rfc/rfc8899#section-3
+//= type=test
+//# Probe loss recovery: It is RECOMMENDED to use probe packets that
+//# do not carry any user data that would require retransmission if
+//# lost.
+
+//= https://www.rfc-editor.org/rfc/rfc8899#section-4.1
+//= type=test
+//# DPLPMTUD MAY choose to use only one of these methods to simplify the
+//# implementation.
+#[test]
+fn on_transmit() {
+    let mut controller = new_controller(1500);
+    controller.state = State::SearchRequested;
+    let now = now();
+    let mut frame_buffer = OutgoingFrameBuffer::new();
+    frame_buffer.set_max_packet_size(Some(controller.probed_size as usize));
+    let mut write_context = MockWriteContext::new(
+        now,
+        &mut frame_buffer,
+        transmission::Constraint::None,
+        transmission::Mode::MtuProbing,
+        endpoint::Type::Server,
+    );
+    let packet_number = write_context.packet_number();
+
+    controller.on_transmit(&mut write_context);
+    assert_eq!(0, write_context.remaining_capacity());
+    assert_eq!(
+        Frame::Ping(frame::Ping),
+        write_context.frame_buffer.pop_front().unwrap().as_frame()
+    );
+    assert_eq!(
+        Frame::Padding(frame::Padding {
+            length: controller.probed_size as usize - 1
+        }),
+        write_context.frame_buffer.pop_front().unwrap().as_frame()
+    );
+    assert_eq!(State::Searching(packet_number, now), controller.state);
+}
